Automatically regenerate parser code during make
The fugly code in Makefile.PL is essentially just generating
a make target for generating the parser code. It's added to the
default action when you just run "make". Adding a new input
format/parser is now as easy as slapping a new Foo.yp file
into the grammars subdirectory and rerunning:

  perl Makefile.PL
  make

There is some small transformation in that the package name
of the generated file is adjusted to read

  Math::ToMath::Parser::$parsername